Refactored `JobManager` into `src/shared`
